fetch-cleared
функції з фетч рознесені по окремих файлах
