+++ src/index.js
@@ -1,7 +1,9 @@
 import './js/fetch';
-import './js/button-scroll-to-top';
-import './js/modal';
-import './js/auth';
-import './js/pagination';
+// import './js/button-scroll-to-top';
+// import './js/modal';
+// import './js/auth';
+// import './js/pagination';
 import './js/loader';
+import './js/render-card-markup';
+import './js/search-film';
